Add new mechanism for DateTime64 binding
