+++ clickhouse_connect/datatypes/numeric.py
@@ -190,7 +190,7 @@ class Bool(ClickHouseType):
         return column
 
     def _write_column_binary(self, column, dest, ctx):
-        write_array('B', [1 if x else 0 for x in column], dest, ctx)
+        write_array('B', [1 if x else 0 for x in column], dest, ctx.column_name)
 
 
 class Boolean(Bool):
@@ -218,15 +218,15 @@ class Enum(ClickHouseType):
         lookup = self._int_map.get
         return [lookup(x, None) for x in column]
 
-    def _write_column_binary(self, column: Union[Sequence, MutableSequence], dest: bytearray, ctx):
+    def _write_column_binary(self, column: Union[Sequence, MutableSequence], dest: bytearray, ctx:InsertContext):
         first = first_value(column, self.nullable)
         if first is None or not isinstance(first, str):
             if self.nullable:
                 column = [0 if not x else x for x in column]
-            write_array(self._array_type, column, dest, ctx)
+            write_array(self._array_type, column, dest, ctx.column_name)
         else:
             lookup = self._name_map.get
-            write_array(self._array_type, [lookup(x, 0) for x in column], dest, ctx)
+            write_array(self._array_type, [lookup(x, 0) for x in column], dest, ctx.column_name)
 
 
 class Enum8(Enum):
@@ -285,15 +285,15 @@ class Decimal(ClickHouseType):
                 app(dec(f'-{digits[:-scale]}.{digits[-scale:]}'))
         return new_col
 
-    def _write_column_binary(self, column: Union[Sequence, MutableSequence], dest: bytearray, _ctx):
-        with decimal.localcontext() as ctx:
-            ctx.prec = self.prec
+    def _write_column_binary(self, column: Union[Sequence, MutableSequence], dest: bytearray, ctx:InsertContext):
+        with decimal.localcontext() as dec_ctx:
+            dec_ctx.prec = self.prec
             dec = decimal.Decimal
             mult = self._mult
             if self.nullable:
-                write_array(self._array_type, [int(dec(str(x)) * mult) if x else 0 for x in column], dest, ctx)
+                write_array(self._array_type, [int(dec(str(x)) * mult) if x else 0 for x in column], dest, ctx.column_name)
             else:
-                write_array(self._array_type, [int(dec(str(x)) * mult) for x in column], dest, ctx)
+                write_array(self._array_type, [int(dec(str(x)) * mult) for x in column], dest, ctx.column_name)
 
     def _active_null(self, ctx: QueryContext):
         if ctx.use_none:
