+++ clickhouse_connect/datatypes/string.py
@@ -104,7 +104,7 @@ class FixedString(ClickHouseType):
                         except UnicodeEncodeError:
                             b = empty
                         if len(b) > sz:
-                            raise ctx.make_data_error(f'UTF-8 encoded FixedString value {b.hex(" ")} exceeds column size {sz}')
+                            raise ctx.data_error(f'UTF-8 encoded FixedString value {b.hex(" ")} exceeds column size {sz}')
                         ext(b)
                         ext(empty[:sz - len(b)])
             else:
@@ -114,7 +114,7 @@ class FixedString(ClickHouseType):
                     except UnicodeEncodeError:
                         b = empty
                     if len(b) > sz:
-                        raise ctx.make_data_error(f'UTF-8 encoded FixedString value {b.hex(" ")} exceeds column size {sz}')
+                        raise ctx.data_error(f'UTF-8 encoded FixedString value {b.hex(" ")} exceeds column size {sz}')
                     ext(b)
                     ext(empty[:sz - len(b)])
         elif self.nullable:
@@ -122,11 +122,11 @@ class FixedString(ClickHouseType):
                 if not b:
                     ext(empty)
                 elif len(b) != sz:
-                    raise ctx.make_data_error(f'Fixed String binary value {b.hex(" ")} does not match column size {sz}')
+                    raise ctx.data_error(f'Fixed String binary value {b.hex(" ")} does not match column size {sz}')
                 else:
                     ext(b)
         else:
             for b in column:
                 if len(b) != sz:
-                    raise ctx.make_data_error(f'Fixed String binary value {b.hex(" ")} does not match column size {sz}')
+                    raise ctx.data_error(f'Fixed String binary value {b.hex(" ")} does not match column size {sz}')
                 ext(b)
